Name two ModelCondition INI callbacks and eight mod-facing sources
The two callbacks are reached only from ModelCondition keys in the exe's own
FieldParse tables (docs/ini_schema.md):

  0x0037D1A0  ExperienceLevel's ModelConditionState, via the 22-entry table at
              0x010EAB48. Single-row file, renamed to match.
  0x0037A9A0  EmotionNugget's ModelConditions and ModelConditionsClear -- one
              callback for two keys because the FieldParse entry supplies the
              store. It parses the list and mirrors the result into the second
              flag set at +0x28.

BfmeConv1410.cpp keeps its name: the other body in it is unrelated and has no
INI reach.

The eight file renames carry no identity claim at all -- each source's ledger
rows already spell a real name and only the filename was address-derived, so a
modder grepping Player, ThingTemplate, SpecialPowerModule, Upgrade or
AttributeModifier now finds the file:

  Rva000C9740Wrap                  -> Player_buildSpecificTeam
  promoted__setCopiedFromDefault_* -> ThingTemplate_setCopiedFromDefault
  Gen002D4C00Destructor            -> CostModifierUpgradeModuleDataDestructor
  promoted__startPowerRecharge_*   -> SpecialPowerModule_startPowerRecharge
  Rva00369860Reset                 -> AttributeModifierDefinition_reset
  Rva00368B80Predicate             -> AttributeModifierPoolUpdate_isModifierActive
  Rva00368C10BitValue              -> AttributeModifierPoolUpdate_getCountForCategoryMask
  Rva000C8CD0DictKnown             -> Dict_known

Landed as one commit because the shared ledger only admits one writer at a
time here and clean windows are scarce.

Also repairs 13 functions.csv rows whose source column had been rolled back to
a path that is in neither HEAD nor the working tree -- a lost update from two
lanes rewriting the shared ledger from stale snapshots. Each was restored to
the source HEAD records for that exact row, which blocked every lane's commit
until now, and lands two sibling renames (0x0024CD10, 0x00685240) that were
finished but stuck behind the same wall.
